+++ HISTORY.rst
@@ -3,6 +3,14 @@
 History
 =======
 
+0.2.0 (2019-07-12)
+---------------------
+
+* Add N50 and N90 metrics.
+* Calculate a simple "Score" for each assembly, the N50/Distance ratio.
+* Combine all metrics in a single table sortable by any column.
+
+
 0.1.0 (2019-06-21)
 ---------------------